MODIFY Update video rendering implementation from upstream
We coped and modfied this implementation from Google's WebRTC implementation.
The original files that inspired this work can be found here:
https://code.google.com/p/webrtc/source/browse/trunk/talk/app/webrtc/objc/

The primary reason we chose this change is because of the assumption that
YUV->RGB pixel conversion can be performed efficiently on the fragment shader,
just in time for rendering. This should free up a small amount of CPU
resources.

Additional features have also been added: Mirroring, disabling the renderer
for temporary loss of video, and frame notifications for further customization.