MODIFY Ensure safe screen capture
Previous implementations were not accounting the mod16 and edge limit
correctly. This should fix both at the same time, by creating a new image of
target size and scaling to fit.

As far as I can tell, this is not possible in a single step call to
drawViewHierarchyInRect, so I'm not even trying.
Closes #117