Added retry mechanism (like recording) for player mode of audio unit
Looks like the iOS 14 compact/full screen mode takes some time to give up audio settings

refer https://stackoverflow.com/questions/26759942/ios-audiounitinitialize-returns-error-code-561017449 for the error we are seeing.
I  turned on the debug flag by default as this is an Audio Driver sample and we better show as many logs as we can, by default
The retry count has been bumped up to 5 from 3 , as losing audio is a big thing in a session, so we try our best.